+++ CHANGELOG.md
@@ -22,6 +22,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added `nix::unistd:seteuid` and `nix::unistd::setegid` for those platforms that do
   not support `setresuid` nor `setresgid` respectively.
   ([#1044](https://github.com/nix-rust/nix/pull/1044))
+- Added a `access` wrapper
+  ([#1045](https://github.com/nix-rust/nix/pull/1045))
 
 ### Changed
 - `PollFd` event flags renamed to `PollFlags` ([#1024](https://github.com/nix-rust/nix/pull/1024/))